+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSAppConfig.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.testkit
+package org.bitcoins.server
 
 import com.typesafe.config.Config
 import org.bitcoins.wallet.config.WalletAppConfig
@@ -6,8 +6,6 @@ import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.chain.config.ChainAppConfig
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
-import java.nio.file.Files
-import com.typesafe.config.ConfigFactory
 
 /**
   * A unified config class for all submodules of Bitcoin-S
@@ -74,62 +72,5 @@ object BitcoinSAppConfig {
   /** Converts the given config to a node config */
   implicit def toNodeConf(conf: BitcoinSAppConfig): NodeAppConfig =
     conf.nodeConf
-
-  /**
-    * App configuration suitable for test purposes:
-    *
-    * 1) Data directory is set to user temp directory
-    */
-  def getTestConfig(config: Config*) = {
-    val tmpDir = Files.createTempDirectory("bitcoin-s-")
-    val confStr = s"""
-    | bitcoin-s {
-    |   datadir = $tmpDir
-    | }
-    |
-    |""".stripMargin
-    val conf = ConfigFactory.parseString(confStr)
-    val allConfs = conf +: config
-    BitcoinSAppConfig(allConfs: _*)
-  }
-
-  sealed trait ProjectType
-
-  object ProjectType {
-    case object Wallet extends ProjectType
-    case object Node extends ProjectType
-    case object Chain extends ProjectType
-
-    val all = List(Wallet, Node, Chain)
-  }
-
-  /** Generates a Typesafe config with DBs set to memory
-    * databases for the given project (or all, if no
-    * project is given). This configuration can then be
-    * given as a override to other configs.
-    */
-  def configWithMemoryDb(project: Option[ProjectType]): Config = {
-    def memConfigForProject(project: ProjectType): String = {
-      val name = project.toString().toLowerCase()
-      s"""
-        | $name.db {
-        |   url = "jdbc:sqlite:file:$name.db:?mode=memory&cache=shared"
-        |   connectionPool = disabled
-        |   keepAliveConnection = true
-        | }
-        |""".stripMargin
-    }
-
-    val confStr = project match {
-      case None    => ProjectType.all.map(memConfigForProject).mkString("\n")
-      case Some(p) => memConfigForProject(p)
-    }
-    val nestedConfStr = s"""
-    | bitcoin-s {
-    | $confStr
-    | }
-    |""".stripMargin
-    ConfigFactory.parseString(nestedConfStr)
-  }
 
 }
